Update JTJ_Disassembler.X68
Displacement should be working. It works fro word for sure but not sure about long. There is not real way to really test it since we are not displacing by more than 32,000 lol. Still need to fix JSR. It calls displacement but its a long every case which its different from the BEQ and BRA instructions so it will need its own displacement thing. Really short, I'll try to do it.