Update JTJ_Disassembler.X68
Finished ADDI and MOVEQ. ADDI works 100% with test code. For some reason MOVEQ would show up as "DATA" when I tried testing. I put a breakpoint in the  CMP.W #$7000,D5  code for MOVEQ opcode and it was comparing 7000 with 7A22, so it went to "DATA" I hard coded in the 7A22 (not on github, obviously)  and MOVEQ was decoded sucessfully.